There was a problem using the PriorityQueue.. I was using a revestres

+++ src/pqueue.py
@@ -3,12 +3,12 @@ from heapq import heappop, heappush
 
 class PriorityQueue:
     def __init__(self):
-        self._pq: list[tuple[int, float]] = []
+        self._pq: list[tuple[float, int]] = []
 
-    def pop(self) -> tuple[int, float]:
+    def pop(self) -> tuple[float, int]:
         return heappop(self._pq)
 
-    def push(self, node: tuple[int, float]):
+    def push(self, node: tuple[float, int]):
         heappush(self._pq, node)
 
     def size(self) -> int:
